--- composeApp/src/commonMain/kotlin/com/rgk/qhatu/common/components/deprecate/AlertDialog.kt
@@ -1,39 +1,0 @@
-package com.rgk.qhatu.components.deprecate
-
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.text.font.FontWeight
-import org.jetbrains.compose.resources.stringResource
-import qhatuapp.composeapp.generated.resources.Res
-import qhatuapp.composeapp.generated.resources.tx_acept
-
-@Composable
-fun AlertMessageDialog(
-    title: String,
-    message: String,
-    confirmText: String = stringResource(Res.string.tx_acept),
-    dismissText: String? = null,
-    onConfirm: () -> Unit,
-    onDismiss: (() -> Unit)? = null
-) {
-    AlertDialog(
-        onDismissRequest = { onDismiss?.invoke() },
-        title = { Text(title, fontWeight = FontWeight.Bold) },
-        text = { Text(message) },
-        confirmButton = {
-            Button(onClick = onConfirm) {
-                Text(confirmText)
-            }
-        },
-        dismissButton = {
-            dismissText?.let {
-                TextButton(onClick = { onDismiss?.invoke() }) {
-                    Text(it)
-                }
-            }
-        }
-    )
-}
