--- composeApp/src/commonMain/kotlin/com/rgk/qhatu/common/components/deprecate/MinimalToolbar.kt
@@ -1,59 +1,0 @@
-package com.rgk.qhatu.components.deprecate
-import androidx.compose.runtime.Composable
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MinimalToolbar(
-    title: String,
-    onClose: () -> Unit,
-    modifier: Modifier = Modifier,
-    backgroundColor: Color = MaterialTheme.colorScheme.primary,
-    contentColor: Color = Color.White
-) {
-    Surface(
-        modifier = modifier.fillMaxWidth(),
-        color = backgroundColor,
-        tonalElevation = 0.dp,
-        shadowElevation = 0.dp
-    ) {
-        TopAppBar(
-            title = {
-                Text(
-                    text = title,
-                    color = contentColor,
-                    style = MaterialTheme.typography.titleMedium
-                )
-            },
-            navigationIcon = {},
-            actions = {
-                IconButton(onClick = onClose) {
-                    Icon(
-                        imageVector = Icons.Default.Close,
-                        contentDescription = "Cerrar",
-                        tint = contentColor
-                    )
-                }
-            },
-            colors = TopAppBarDefaults.topAppBarColors(
-                containerColor = backgroundColor,
-                titleContentColor = contentColor,
-                actionIconContentColor = contentColor
-            )
-        )
-    }
-}
