+++ composeApp/src/commonMain/kotlin/com/rgk/qhatu/feature/sale/presentation/receipt/searchprovider/SearchProviderScreen.kt
@@ -8,24 +8,14 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBackIosNew
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.rgk.qhatu.components.deprecate.AppToolbar
-import com.rgk.qhatu.components.deprecate.ErrorView
-import com.rgk.qhatu.components.deprecate.LoadingView
-import org.jetbrains.compose.resources.stringResource
-import qhatuapp.composeapp.generated.resources.Res
-import qhatuapp.composeapp.generated.resources.tx_back
+import com.rgk.qhatu.common.components.loading.LoadingView
 
 @Composable
 fun SearchProviderScreen(
@@ -36,17 +26,6 @@ fun SearchProviderScreen(
 
     Box(modifier = Modifier.fillMaxSize()) {
         Column(modifier = Modifier.fillMaxSize()) {
-            AppToolbar(
-                title = "Search",
-                navigationIcon = {
-                    IconButton(onClick = { navController.popBackStack() }) {
-                        Icon(
-                            imageVector = Icons.Default.ArrowBackIosNew,
-                            contentDescription = stringResource(Res.string.tx_back)
-                        )
-                    }
-                }
-            )
 
             if (uiState is SearchProviderUiState.Success) {
                 LazyColumn(
@@ -80,12 +59,7 @@ fun SearchProviderScreen(
         }
 
         if (uiState is SearchProviderUiState.Error) {
-            ErrorView(
-                message = uiState.message,
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .padding(16.dp)
-            )
+
         }
     }
