+++ composeApp/src/commonMain/kotlin/com/rgk/qhatu/feature/setting/presentation/store/StoreScreen.kt
@@ -12,8 +12,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.rgk.qhatu.components.deprecate.CustomTextField
-import com.rgk.qhatu.components.deprecate.CustomTextFieldParams
+import com.rgk.qhatu.common.components.textfield.CustomTextField
+import com.rgk.qhatu.common.components.textfield.CustomTextFieldParams
 import com.rgk.qhatu.feature.setting.domain.model.Store
 import org.jetbrains.compose.resources.stringResource
 import qhatuapp.composeapp.generated.resources.Res
